requirements.txt
streamlit
google.generativeai
streamlit_option_menu
google.cloud
google.oauth2.service_account